moving prettier to scripts package
sadly, no version of it can run in the browser
