+++ packages/sdk-codegen-scripts/src/index.ts
@@ -31,3 +31,4 @@ export * from './nodeUtils'
 export * from './reformatter'
 export * from './sdkConfig'
 export * from './sdkGenerator'
+export * from './prettify'
