+++ packages/sdk-codegen-scripts/src/prettify.ts
