+++ packages/sdk-codegen/package.json
@@ -38,12 +38,10 @@
     "@looker/sdk-codegen-utils": "^21.0.11",
     "@looker/sdk-rtl": "^21.0.20",
     "blueimp-md5": "^2.13.0",
-    "openapi3-ts": "^1.3.0",
-    "prettier": "^2.4.1"
+    "openapi3-ts": "^1.3.0"
   },
   "gitHead": "308cb8c61290837e71345a5ef6fb17be9285990e",
   "devDependencies": {
-    "@types/prettier": "^2.3.2",
     "lodash": "^4.17.20"
   }
 }
