+++ packages/sdk-codegen/src/index.ts
@@ -37,4 +37,3 @@ export * from './exampleInfo'
 export * from './declarationInfo'
 export * from './specDiff'
 export * from './specConverter'
-export * from './prettify'
